requirements.txt
# dogArm Project Python Requirements

# Core dependencies
numpy>=1.21.0
pyserial>=3.5
matplotlib>=3.5.0

# For calligraphy path planning
opencv-python>=4.5.0
Pillow>=9.0.0

# For data processing
pandas>=1.3.0

# Optional: for GUI (if needed)
# PyQt5>=5.15.0
